[DAPS-14xx] Move files over and move out svg
