+++ web/static/components/provenance/panel_graph.js
@@ -1,7 +1,12 @@
-import * as util from "./util.js";
-import * as model from "./model.js";
-import * as api from "./api.js";
-import * as panel_info from "./panel_item_info.js";
+import * as util from "../../util.js";
+import * as model from "../../model.js";
+import * as api from "../../api.js";
+import * as panel_info from "../../panel_item_info.js";
+import {
+    defineArrowMarkerComp,
+    defineArrowMarkerDeriv,
+    defineArrowMarkerNewVer,
+} from "./assets/arrow-markers.js";
 
 export function newGraphPanel(a_id, a_frame, a_parent) {
     return new GraphPanel(a_id, a_frame, a_parent);
@@ -874,66 +879,4 @@ function graphPruneCalc(a_node, a_visited, a_source) {
     }
 
     return a_node.prune;
-}
-
-function defineArrowMarkerDeriv(a_svg) {
-    a_svg
-        .append("defs")
-        .append("marker")
-        .attr("id", "arrow-derivation")
-        .attr("refX", -2.5)
-        .attr("refY", 2)
-        .attr("orient", "auto")
-        .attr("markerWidth", 5)
-        .attr("markerHeight", 4)
-        .append("svg:path")
-        .attr("class", "arrow-path derivation")
-        .attr("d", "M 5,0 L 0,2 L 5,4");
-}
-
-function defineArrowMarkerComp(a_svg) {
-    a_svg
-        .append("defs")
-        .append("marker")
-        .attr("id", "arrow-component")
-        .attr("refX", -2.5)
-        .attr("refY", 2)
-        .attr("orient", "auto")
-        .attr("markerWidth", 8)
-        .attr("markerHeight", 4)
-        .append("svg:path")
-        .attr("class", "arrow-path component")
-        .attr("d", "M 4,0 L 0,2 L 4,4 L 8,2");
-}
-
-/*
-Old version marker at 'start'
-function defineArrowMarkerNewVer( a_svg, a_name ){
-    a_svg.append('defs').append('marker')
-        .attr('id','arrow-new-version')
-        .attr('refX',8.5)
-        .attr('refY',2)
-        .attr('orient','auto')
-        .attr('markerWidth',10)
-        .attr('markerHeight',4)
-        .append('svg:path')
-            .attr('class','arrow-path new-version')
-            .attr('d', 'M 2,0 L 6,2 L 2,4 M 4,2 L 0,4 L 0,0');
-}
-*/
-
-// New version marker at 'end'
-function defineArrowMarkerNewVer(a_svg, a_name) {
-    a_svg
-        .append("defs")
-        .append("marker")
-        .attr("id", "arrow-new-version")
-        .attr("refX", -2.5)
-        .attr("refY", 2)
-        .attr("orient", "auto")
-        .attr("markerWidth", 7)
-        .attr("markerHeight", 4)
-        .append("svg:path")
-        .attr("class", "arrow-path new-version")
-        .attr("d", "M 5,0 L 0,2 L 5,4 L 5,0 M 6,0 L 7,0 L 7,4 L 6,4 L 6,0");
 }
