+++ web/views/main.ect
@@ -50,7 +50,7 @@
         <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/dlg_schema_list.js"></script>
         <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/dlg_query_builder.js"></script>
         <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/panel_catalog.js"></script>
-        <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/panel_graph.js"></script>
+        <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/components/provenance/panel_graph.js"></script>
         <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/panel_item_info.js"></script>
         <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/main_browse_tab.js"></script>
         <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/main.js"></script>
